Skyline: promote WaitForCondition from AbstractFunctionalTest to AbstractUnitTest in an effort to avoid an apparent race condition in ConsoleBadRawFileImportTest, which only seems to fail under code coverage testing.
